+++ ts/packages/knowPro/src/conversationIndex.ts
@@ -27,11 +27,6 @@ import { IConversationSecondaryIndexes } from "./secondaryIndexes.js";
 import { addPropertiesToIndex, PropertyIndex } from "./propertyIndex.js";
 import { TimestampToTextRangeIndex } from "./timestampIndex.js";
 
-export interface IConversationIndexes {
-    semanticRefIndex?: ITermToSemanticRefIndex | undefined;
-    secondaryIndexes?: IConversationSecondaryIndexes | undefined;
-}
-
 export function createKnowledgeModel() {
     const chatModelSettings = openai.apiSettingsFromEnv(
         openai.ModelType.Chat,
